whisper/requirements.txt
aiohttp==3.6.3
aiohttp-basicauth==1.0.0
aiohttp-cors==0.6.0
async-timeout==3.0.1
attrs==23.1.0
av==10.0.0
azure-cognitiveservices-speech==1.29.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==3.15.0
filelock==3.12.0
flatbuffers==23.5.26
fsspec==2023.5.0
furl==2.1.3
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
mpmath==1.3.0
multidict==4.7.6
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.8
orderedmultidict==1.0.1
packaging==23.1
protobuf==4.23.2
pyreadline3==3.4.1
python-engineio==3.13.2
python-socketio==4.6.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
sympy==1.12
tokenizers==0.13.3
tqdm==4.65.0
typing-extensions==4.6.3
urllib3==2.0.2

